+++ source/others/i64_to_f32.cpp
@@ -41,7 +41,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 float32_t i64_to_f32(int64_t a)
 {
     bool const sign = (a < 0);
-    uint64_t const absA = sign ? -a : a;
+    uint64_t const absA = static_cast<uint64_t>(sign ? -a : a);
     int8_t shiftDist = softfloat_countLeadingZeros64(absA) - 40;
     if (0 <= shiftDist) {
         return u_as_f_32(a ? packToF32UI(sign, 0x95 - shiftDist, (uint32_t)absA << shiftDist) : 0);
